requirements.txt
streamlit>=1.22.0
numpy>=1.20.0
matplotlib>=3.5.0
python-dotenv>=0.19.0

# Dependencies for AI model
google-generativeai>=0.8.0  # Google Gemini API
